+++ doc/source/changelog.rst
@@ -6,6 +6,8 @@ next
 * Added support for Python 3.10.
 * Fixed bad interaction between custom None parsers and tuple parsers.
 * Give preference to None parsers in Unions.
+* `defopt.signature` annotates the return type with the documented raisable
+  exception types.
 
 6.1.0 (2021-02-25)
 ------------------